+++ lib/src/logger/result_error.dart
@@ -1,0 +1,6 @@
+part of tie_fp;
+
+abstract class ResultError {
+  static bool enableLog = false;
+  static ValueChanged<Failure>? onError;
+}