+++ lib/src/tie_fp.dart
@@ -1,0 +1,9 @@
+library tie_fp;
+
+import 'package:flutter/cupertino.dart';
+
+part 'extensions/extensions.dart';
+part 'logger/result_error.dart';
+part 'result/failure.dart';
+part 'result/result.dart';
+part 'result/success.dart';