+++ lib/tie_fp.dart
@@ -1,4 +1,9 @@
 library tie_fp;
 
-part 'src/extensions.dart';
-part 'src/result.dart';
+import 'package:flutter/cupertino.dart';
+
+part 'src/extensions/extensions.dart';
+part 'src/logger/result_error.dart';
+part 'src/result/failure.dart';
+part 'src/result/result.dart';
+part 'src/result/success.dart';