+++ test/tie_fp_test.dart
@@ -1,16 +1,40 @@
-// import 'package:tie_fp/tie_fp.dart';
-// import 'package:test/test.dart';
+import 'package:test/test.dart';
+import 'package:tie_fp/tie_fp.dart';
 
-// void main() {
-//   group('A group of tests', () {
-//     final awesome = Awesome();
+void main() {
+  group('Result test', () {
+    late Result value;
+    late Result valueError;
+    setUp(() {
+      value = Success(1);
+      valueError = Failure(Exception('custom message'));
+      return;
+    });
 
-//     setUp(() {
-//       // Additional setup goes here.
-//     });
-
-//     test('First Test', () {
-//       expect(awesome.isAwesome, isTrue);
-//     });
-//   });
-// }
+    test('Value returns no error', () {
+      expect(value.isError(), false);
+    });
+    test('Value is int', () {
+      expect(value.getValue(), isA<int?>());
+    });
+    test('Value is 1', () {
+      expect(value.getValue(), 1);
+    });
+    test(
+      'Value Error is error',
+      () => expect(valueError.isError(), true),
+    );
+    test(
+      'Value Error get error is exception',
+      () => expect(valueError.getRawError(), isException),
+    );
+    test(
+      'Value Error value is null',
+      () => expect(valueError.getValue(), null),
+    );
+    test(
+      'Value Error value is null',
+      () => expect(valueError.getValue(), null),
+    );
+  });
+}